task1/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.31
qdrant-client==1.15.1
python-dotenv==1.1.1
openai==1.101.0
